fix(ci): unblock dashboard trend checks
Split the trend panel out of DashboardOverview to stay within frontend source budgets.

Retry Xray release downloads in Docker builds to tolerate transient GitHub 504 failures.
