+++ Dockerfile
@@ -38,7 +38,9 @@ RUN apt-get update \
         arm64) XRAY_ZIP="Xray-linux-arm64-v8a.zip" ;; \
         *) echo "Unsupported TARGETARCH=${TARGETARCH} resolved_arch=${ARCH} for Xray-core" >&2; exit 1 ;; \
       esac \
-    && curl -fsSL -o /tmp/xray.zip "https://github.com/XTLS/Xray-core/releases/download/v${XRAY_CORE_VERSION}/${XRAY_ZIP}" \
+    && curl --fail --show-error --silent --location \
+      --retry 5 --retry-delay 2 --retry-all-errors \
+      -o /tmp/xray.zip "https://github.com/XTLS/Xray-core/releases/download/v${XRAY_CORE_VERSION}/${XRAY_ZIP}" \
     && unzip -q /tmp/xray.zip -d /tmp/xray \
     && install -m 0755 /tmp/xray/xray /usr/local/bin/xray \
     && install -d /usr/local/share/licenses/xray-core \
